requirements.txt
requests
beautifulsoup4
wikipedia
python-dotenv